fix:node to suite protobuf

+++ src/com/intel/fangpei/BasicMessage/BasicMessage.java
@@ -9,21 +9,21 @@ import java.util.Map;
  */
 public class BasicMessage {
 	Map<Integer, String> s = new HashMap<Integer, String>();
-	public static final byte NODE = (byte) 81;
-	public static final byte ADMIN = (byte) 80;
-	public static final byte SERVER = (byte) 79;
-	public static final byte OP_LOGIN = (byte) 15;
-	public static final byte OP_QUIT = (byte) 10;
-	public static final byte OP_CLOSE = (byte) 9;
-	public static final byte OP_EXEC = (byte) 1;
-	public static final byte OP_MESSAGE = (byte) 2;
-	public static final byte OP_SYSINFO = (byte) 11;
-//	public static final byte OP_lOAD_DISK = (byte) 5;
+	public static final int NODE = 81;
+	public static final int ADMIN =   80;
+	public static final int SERVER =   79;
+	public static final int OP_LOGIN =   15;
+	public static final int OP_QUIT =   10;
+	public static final int OP_CLOSE =   9;
+	public static final int OP_EXEC =   1;
+	public static final int OP_MESSAGE =   2;
+	public static final int OP_SYSINFO =   11;
+//	public static final byte OP_lOAD_DISK =   5;
 	public static final int VERSION = 1;
-	public static final byte OP_HELP = (byte) 8;
-	public static final byte OK = (byte) 7;
-	public static final byte OP_SH = (byte) 12;
-	public static final byte TASKS = (byte) 31;
-	public static final byte TASKINFO = (byte) 32;
-	public static final byte CHILDINFO = (byte)33;
+	public static final int OP_HELP =   8;
+	public static final int OK =   7;
+	public static final int OP_SH =   12;
+	public static final int TASKS =   31;
+	public static final int TASKINFO =   32;
+	public static final int CHILDINFO =  33;
 }
